pyproject.toml
[project]
name = "comfyui-firefly-services"
version = "2.0.0"
description = "ComfyUI custom nodes for Adobe Firefly, Photoshop, Substance 3D, InDesign, Video Reframe, and utility operations"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8.0",
    "pydantic>=2.0.0",
    "boto3>=1.28.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "psd-tools>=1.9.0",
]

[project.urls]
Repository = "https://github.com/crehop/firefly_services_nodes"

[tool.comfyui]
PublisherId = "crehop"
DisplayName = "Firefly Services"
Icon = ""
